Add custom minute span IT

+++ core/src/main/java/org/opensearch/sql/calcite/udf/SpanFunction.java
@@ -7,6 +7,7 @@ package org.opensearch.sql.calcite.udf;
 
 import java.time.LocalDate;
 import java.time.ZoneOffset;
+import org.apache.calcite.linq4j.function.Strict;
 import org.apache.calcite.runtime.SqlFunctions;
 import org.apache.calcite.sql.type.SqlTypeName;
 import org.opensearch.sql.planner.physical.collector.Rounding.DateTimeUnit;
@@ -42,14 +43,14 @@ public class SpanFunction implements UserDefinedFunction {
 
     switch (sqlTypeName) {
       case SqlTypeName.DATE:
-        LocalDate date = LocalDate.ofEpochDay((int) args[0]);
+        LocalDate date = LocalDate.ofEpochDay(((Integer) args[0]).longValue());
         long dateEpochValue =
             dateTimeUnit.round(
                 date.atStartOfDay().atZone(ZoneOffset.UTC).toInstant().toEpochMilli(), interval);
         return SqlFunctions.timestampToDate(dateEpochValue);
       case SqlTypeName.TIME:
         /*
-         * Follow current logic to ignore time frame greater than hour because TIME type like '17:59:59' doesn't have day, month, year, etc.
+         * Follow current logic to ignore time frame greater than hour because TIME type like '17:59:59.99' doesn't have day, month, year, etc.
          * See @org.opensearch.sql.planner.physical.collector.TimeRounding
          */
         if (dateTimeUnit.getId() > 4) {
